requirements.txt
streamlit
anthropic
agno
PyPDF2
python-dotenv
